LayoutTests/imported/w3c/web-platform-tests/svg/animations/keysplines-y-limits-expected.txt

CONSOLE MESSAGE: Error: Invalid value for <animate> attribute keySplines="0 -1 1 1"
CONSOLE MESSAGE: Error: Invalid value for <animate> attribute keySplines="0 2 1 1"
CONSOLE MESSAGE: Error: Invalid value for <animate> attribute keySplines="0 0 1 -1"
CONSOLE MESSAGE: Error: Invalid value for <animate> attribute keySplines="0 0 1 2"
CONSOLE MESSAGE: Error: Invalid value for <animateMotion> attribute keySplines="0 -10 1 1"
CONSOLE MESSAGE: Error: Invalid value for <animateMotion> attribute keySplines="0 10 1 1"
CONSOLE MESSAGE: Error: Invalid value for <animateMotion> attribute keySplines="0 0 1 -10"
CONSOLE MESSAGE: Error: Invalid value for <animateMotion> attribute keySplines="0 0 1 10"

PASS 'keySplines' with y-values outside of the 0 to 1 range
